move files into folders
