+++ src/components/App/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import './App.css';
-import { fetchMovies } from './apiCalls';
+import NavBar from '../NavBar/NavBar';
+import { fetchMovies } from '../../api/apiCalls';
 
 class App extends Component {
   constructor() {
@@ -18,8 +19,7 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        <header className="App-header">
-        </header>
+        <NavBar />
       </div>
     );
   }
